Finished Find Route , Added Email verification on Profile Page, Some Improvments added

+++ WolfTaxi_WPF/MVVM/ViewModels/ProfilePageVM.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
 using System.Runtime.CompilerServices;
+using System.Security;
 using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
@@ -10,10 +11,14 @@ using System.Windows;
 using System.Windows.Media;
 using Microsoft.Win32;
 using WolfTaxi_WPF.Commands;
+using WolfTaxi_WPF.Enums;
 using WolfTaxi_WPF.MVVM.Models.DerivedClasses;
 using WolfTaxi_WPF.MVVM.Views;
 using WolfTaxi_WPF.MVVM.Views.Pages;
 using WolfTaxi_WPF.Services;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement.ListView;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement.StartPanel;
+using static System.Windows.Forms.VisualStyles.VisualStyleElement.TaskbarClock;
 
 namespace WolfTaxi_WPF.MVVM.ViewModels
 {
@@ -26,6 +31,8 @@ namespace WolfTaxi_WPF.MVVM.ViewModels
 
         public string ToolTipStr { get; set; } = "Password Requirements :\n* Use at least 8 characters \n* Use upper and lower case characters \n* Use 1 or more numbers \n* Recommend use special characters";
         public bool IsEditing { get; set; } = false;
+        private int SecurityCode { get; set; }
+
 
         public ProfilePage Page { get; set; }
         public string UrlTempPP { get; set; } = string.Empty;
@@ -41,6 +48,28 @@ namespace WolfTaxi_WPF.MVVM.ViewModels
         #endregion
 
         #region Methods
+
+        public bool SendCode()
+        {
+            if (User.Email == App.DataFacade.User.Email)
+                return true;
+
+            SendMail();
+            EnterSecurityCode enter = new();
+            enter.Reset();
+            enter.Code = SecurityCode;
+            enter.ShowDialog();
+            if (enter.DialogResult != DialogResult.Success)
+                return false;
+
+            return true;
+        }
+
+        public void SendMail()
+        {
+            SecurityCode = new Random().Next(1000, 9999);
+            EmailService.SendSecurityCode(User.Email, "Email OTP: ", SecurityCode.ToString(), "WolfTaxi");
+        }
 
         public void Browse(object param)
         {
@@ -64,6 +93,15 @@ namespace WolfTaxi_WPF.MVVM.ViewModels
 
         public void SaveRun(object param)
         {
+            if (!SendCode())
+            {
+                CMessageBox.Show($"{User.Email} Not Verified !", CMessageTitle.Error, CMessageButton.Ok, CMessageButton.None);
+                Page.EmailIsTrue = false;
+                return;
+            }
+
+            Page.EmailIsTrue = true;
+
             if (dialogResult)
                 CloudinaryService.DestroyImage("tempdriverpp", App.TempCloudinaryFolderPath);
             User.SourceOfPP = dialogResult ? CloudinaryService.UploadImage(ofd.FileName, User.Username, App.UserCloudinaryFolderPath) : App.DriverProfilePhoto;
@@ -86,6 +124,8 @@ namespace WolfTaxi_WPF.MVVM.ViewModels
         public RelayCommand EditCommand { get; set; }
 
         public RelayCommand ShowPass { get; set; }
+
+        public RelayCommand EditEmail { get; set; }
 
         public RelayCommand Save { get; set; }
